Refactors day 03
