+++ lib/geometry/geometry.py
